refactor(suite): rename extra dependencies package
