+++ packages/suite/package.json
@@ -49,6 +49,7 @@
         "@suite-common/earn-stablecoin": "workspace:^",
         "@suite-common/earn-stablecoin-api": "workspace:*",
         "@suite-common/earn-staking-api": "workspace:*",
+        "@suite-common/extra-dependencies": "workspace:*",
         "@suite-common/feedback": "workspace:*",
         "@suite-common/fiat-services": "workspace:*",
         "@suite-common/firmware": "workspace:*",
@@ -65,7 +66,6 @@
         "@suite-common/platform-encryption": "workspace:*",
         "@suite-common/react-query": "workspace:*",
         "@suite-common/receive": "workspace:*",
-        "@suite-common/redux-extra-dependencies": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
         "@suite-common/sentry": "workspace:*",
         "@suite-common/staking": "workspace:*",
