+++ suite-common/extra-dependencies/forbiddenDeps.config.ts
@@ -2,14 +2,9 @@ import type { ForbiddenDepsConfig } from '@trezor/requirements';
 
 export const forbiddenDepsConfig: ForbiddenDepsConfig = {
     'allowed-only-in': {
-        packages: [
-            '@trezor/suite',
-            '@suite-common/test-utils',
-            '@suite-native/state',
-            '@suite-native/test-utils',
-        ],
+        packages: ['@trezor/suite', '@suite-native/state'],
         reason:
             'The global dependency graph may be imported only by desktop and Native application ' +
-            'wiring and their dedicated dependency mocks.',
+            'wiring.',
     },
 };
