+++ suite-common/extra-dependencies/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@suite-common/redux-extra-dependencies",
+    "name": "@suite-common/extra-dependencies",
     "version": "1.0.0",
     "private": true,
     "license": "See LICENSE.md in repo root",
