+++ suite-common/extra-dependencies/src/index.ts
@@ -1,0 +1,3 @@
+export type * from './CommonServices';
+export type * from './ExtraDependenciesStatic';
+export * from './notImplemented';
