--- suite-common/redux-extra-dependencies/src/index.ts
@@ -1,2 +1,0 @@
-export type * from './extraDependenciesType';
-export * from './notImplemented';
