+++ suite-native/state/package.json
@@ -20,6 +20,7 @@
         "@suite-common/dependency-injection": "workspace:*",
         "@suite-common/device": "workspace:*",
         "@suite-common/discreet-mode": "workspace:*",
+        "@suite-common/extra-dependencies": "workspace:*",
         "@suite-common/firmware": "workspace:*",
         "@suite-common/geolocation": "workspace:*",
         "@suite-common/logger": "workspace:*",
@@ -27,7 +28,6 @@
         "@suite-common/networks": "workspace:*",
         "@suite-common/platform-encryption-native": "workspace:*",
         "@suite-common/receive": "workspace:*",
-        "@suite-common/redux-extra-dependencies": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
         "@suite-common/suite-rbf-labels-migrations": "workspace:*",
         "@suite-common/suite-sync": "workspace:*",
